feat(client): drop deprecated middleware feature
