+++ packages/client-generator-js/src/TSClient/PrismaClient.ts
@@ -484,13 +484,6 @@ export class PrismaClient<
    */
   $disconnect(): $Utils.JsPromise<void>;
 
-  /**
-   * Add a middleware
-   * @deprecated since 4.16.0. For new code, prefer client extensions instead.
-   * @see https://pris.ly/d/extensions
-   */
-  $use(cb: Prisma.Middleware): void
-
 ${[
   executeRawDefinition(this.context),
   queryRawDefinition(this.context),
