+++ packages/client-generator-ts/src/TSClient/PrismaClient.ts
@@ -318,13 +318,6 @@ export interface PrismaClient<
    */
   $disconnect(): runtime.Types.Utils.JsPromise<void>;
 
-  /**
-   * Add a middleware
-   * @deprecated since 4.16.0. For new code, prefer client extensions instead.
-   * @see https://pris.ly/d/extensions
-   */
-  $use(cb: Prisma.Middleware): void
-
 ${[
   executeRawDefinition(this.context),
   queryRawDefinition(this.context),
