--- packages/client/src/__tests__/integration/happy/async-hooks/.gitignore
@@ -1,1 +1,0 @@
-!dev.db
