--- packages/client/src/__tests__/integration/happy/async-hooks/schema.prisma
@@ -1,28 +1,0 @@
-datasource db {
-  provider = "sqlite"
-  url      = "file:dev.db"
-}
-
-generator client {
-  provider             = "prisma-client-js"
-}
-
-// / User model comment
-model User {
-  id    String  @default(uuid()) @id
-  email String  @unique
-  // / name comment
-  name  String?
-  posts Post[]
-}
-
-model Post {
-  id        String   @default(cuid()) @id
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-  published Boolean
-  title     String
-  content   String?
-  authorId  String?  @map("author")
-  author    User?    @relation(fields: [authorId], references: [id])
-}
