--- packages/client/src/__tests__/integration/happy/middlewares-manipulation/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "my-prisma-project",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
