--- packages/client/src/__tests__/integration/happy/middlewares-transaction/.gitignore
@@ -1,1 +1,0 @@
-!dev.db
