--- packages/client/src/__tests__/integration/happy/middlewares-transaction/schema.prisma
@@ -1,28 +1,0 @@
-datasource db {
-  provider = "sqlite"
-  url      = "file:dev.db"
-}
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-// / User model comment
-model User {
-  id    String  @id @default(uuid())
-  email String  @unique
-  // / name comment
-  name  String?
-  posts Post[]
-}
-
-model Post {
-  id        String   @id @default(cuid())
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-  published Boolean
-  title     String
-  content   String?
-  authorId  String?
-  author    User?    @relation(fields: [authorId], references: [id])
-}
