--- packages/client/src/__tests__/integration/happy/middlewares/.gitignore
@@ -1,1 +1,0 @@
-!dev.db
