--- packages/client/src/__tests__/types/middlewares/index.d.ts
@@ -1,1 +1,0 @@
-export { PrismaClient } from '@prisma/client'
