--- packages/client/src/__tests__/types/middlewares/index.test-d.ts
@@ -1,15 +1,0 @@
-import { expectError } from 'tsd'
-
-import { PrismaClient } from '.'
-
-const prisma = new PrismaClient({
-  datasources: {
-    db: {
-      url: 'file:dev.db',
-    },
-  },
-})
-
-;(async () => {
-  expectError(prisma.$use('event'))
-})()
