--- packages/client/src/__tests__/types/middlewares/test.ts
@@ -1,19 +1,0 @@
-import { PrismaClient } from '@prisma/client'
-
-// This file will not be executed, just compiled to check if the typings are valid
-async function main() {
-  const prisma = new PrismaClient()
-
-  prisma.$use(async (params, next) => {
-    params.args
-    params.dataPath
-    params.action
-    params.model as 'user' | 'Post' | 'MachineData' | 'Like' | undefined
-    params.runInTransaction
-    return next(params)
-  })
-}
-
-main().catch((e) => {
-  console.error(e)
-})
