--- packages/client/src/__tests__/types/middlewares/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2021",
-    "module": "commonjs",
-    "strict": true,
-    "esModuleInterop": true
-  }
-}
