+++ packages/client/src/runtime/core/extensions/$extends.ts
@@ -19,7 +19,6 @@ export function $extends(this: Client, extension: ExtensionArgs | ((client: Clie
   const newClient = Object.create(this._originalClient, {
     _extensions: { value: this._extensions.append(extension) },
     _appliedParent: { value: this, configurable: true },
-    $use: { value: undefined },
     $on: { value: undefined },
   }) as Client
