--- packages/client/tests/functional/issues/15644-middleware-arg-types/_matrix.ts
@@ -1,13 +1,0 @@
-import { defineMatrix } from '../../_utils/defineMatrix'
-import { Providers } from '../../_utils/providers'
-
-export default defineMatrix(() => [
-  [
-    {
-      provider: Providers.SQLITE,
-    },
-    {
-      provider: Providers.MONGODB,
-    },
-  ],
-])
