--- packages/client/tests/functional/middleware-raw-args/_matrix.ts
@@ -1,4 +1,0 @@
-import { defineMatrix } from '../_utils/defineMatrix'
-import { sqlProviders } from '../_utils/providers'
-
-export default defineMatrix(() => [sqlProviders])
