--- packages/client/tests/functional/middleware-raw-args/prisma/_schema.ts
@@ -1,20 +1,0 @@
-import { idForProvider } from '../../_utils/idForProvider'
-import testMatrix from '../_matrix'
-
-export default testMatrix.setupSchema(({ provider }) => {
-  return /* Prisma */ `
-  generator client {
-    provider = "prisma-client-js"
-  }
-  
-  datasource db {
-    provider = "${provider}"
-    url      = env("DATABASE_URI_${provider}")
-  }
-  
-  model Entry {
-    id ${idForProvider(provider)}
-    name String
-  }
-  `
-})
